add:gacha_record_part_1
